Clean up GuiUi class
