+++ src/seedu/addressbook/Main.java
@@ -9,26 +9,24 @@ import seedu.addressbook.ui.GuiUi;
 
 public class Main extends Application {
 
-
-    private GuiUi ui;
-    private Logic logic;
+    private GuiUi gui;
 
     @Override
     public void start(Stage primaryStage) throws Exception{
-        logic = new Logic();
-        ui = new GuiUi(logic);
-        ui.start(primaryStage);
-    }
-
-    public static void main(String[] args) {
-        launch(args);
+        Logic logic = new Logic();
+        gui = new GuiUi(logic);
+        gui.start(primaryStage);
     }
 
     @Override
     public void stop() throws Exception {
         super.stop();
-        ui.showGoodbyeMessage();
+        gui.stop();
         Platform.exit();
         System.exit(0);
+    }
+
+    public static void main(String[] args) {
+        launch(args);
     }
 }
